Bugfixes sezz (#534)
* Fix #500: "Throwing the torch while running will cause it to fall behind Lara"; make flare and torch dynamics control functions more similar; cleanup

* Fix #473: "Sound: Harpoon gun does not make shoot sfx for the 4th bolt"

* Possibly fix #492: "Landing on teeth spikes from falling not working as expected (either stuck in mid-air, or wrong animation)"

* Cleanup

* Update tr4_mutant.cpp and tr4_wild_boar.cpp

* Update tr4_sphinx.cpp

* Fix typo

* Fix torches and flares sliding around too much

* Tweak minimum jump splat speed

* Restore snap to height upon starting a slide

* Possibly fix minecart dismounts

* Fix merge error

* Simplified translation (#520)

* Simplify translation of items and vectors
* Update lara_one_gun.cpp
* Restore comments
* Reorder functions
* Restore a GetCollision() overload
* Rename stuff and upset everyone
* Rename many, many, many variables
* Add TranslateVector() overloads
* Simplify
* Update translation functions
* Rename
* Update trmath.cpp
* Fix typo
* Update trmath.cpp


* Use GetLaraInfo(); cleanup

* Minor flame_emitters.cpp cleanup

* Fix object ID error and convert to switch case

* Fix edge case of skewed movement angle on slopes; cleanup

* Prevent crouch roll while using binoculars

* TouchBits cleanup

* TR4 entity cleanup

* TR4 entity cleanup

* Partially fill out enums in tr4_skeleton.cpp; cleanup

* TouchJoints cleanup

+++ TombEngine/Game/Lara/lara_fire.h
@@ -22,12 +22,12 @@ struct WeaponInfo
 	short ShotAccuracy;
 	int GunHeight;
 	short TargetDist;
-	byte Damage;
-	byte RecoilFrame;
-	byte FlashTime;
-	byte DrawFrame;
+	int Damage;
+	int RecoilFrame;
+	int FlashTime;
+	int DrawFrame;
 	short SampleNum;
-	byte ExplosiveDamage;
+	int ExplosiveDamage;
 };
 
 enum WeaponState
